refactor: move core extensions to separate package.
